Support for custom video frame renderers. Default implementation maintains current "scale to fit" rendering.
