+++ litr/src/main/java/com/linkedin/android/litr/render/GlVideoRenderer.java
@@ -21,20 +21,20 @@
 package com.linkedin.android.litr.render;
 
 import android.media.MediaFormat;
-import android.opengl.GLES11Ext;
 import android.opengl.GLES20;
 import android.opengl.Matrix;
 import android.os.Build;
 import android.view.Surface;
+
 import androidx.annotation.Nullable;
+
 import com.linkedin.android.litr.codec.Frame;
 import com.linkedin.android.litr.filter.GlFilter;
+import com.linkedin.android.litr.filter.GlFrameRenderFilter;
+import com.linkedin.android.litr.filter.video.gl.ScaleToFitGlFrameRenderFilter;
 
-import java.nio.ByteBuffer;
-import java.nio.ByteOrder;
-import java.nio.FloatBuffer;
+import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 /**
@@ -47,65 +47,37 @@ public class GlVideoRenderer implements Renderer {
                                                  ? MediaFormat.KEY_ROTATION
                                                  : "rotation-degrees";
 
-    private static final int FLOAT_SIZE_BYTES = 4;
-    private static final int TRIANGLE_VERTICES_DATA_STRIDE_BYTES = 5 * FLOAT_SIZE_BYTES;
-    private static final int TRIANGLE_VERTICES_DATA_POS_OFFSET = 0;
-    private static final int TRIANGLE_VERTICES_DATA_UV_OFFSET = 3;
-
-    // shaders
-    private static final String VERTEX_SHADER =
-        "uniform mat4 uMVPMatrix;\n" +
-            "uniform mat4 uSTMatrix;\n" +
-            "attribute vec4 aPosition;\n" +
-            "attribute vec4 aTextureCoord;\n" +
-            "varying vec2 vTextureCoord;\n" +
-            "void main() {\n" +
-            "  gl_Position = uMVPMatrix * aPosition;\n" +
-            "  vTextureCoord = (uSTMatrix * aTextureCoord).xy;\n" +
-            "}\n";
-    private static final String FRAGMENT_SHADER =
-        "#extension GL_OES_EGL_image_external : require\n" +
-            "precision mediump float;\n" +      // highp here doesn't seem to matter
-            "varying vec2 vTextureCoord;\n" +
-            "uniform samplerExternalOES sTexture;\n" +
-            "void main() {\n" +
-            "  gl_FragColor = texture2D(sTexture, vTextureCoord);\n" +
-            "}\n";
-
     private VideoRenderInputSurface inputSurface;
     private VideoRenderOutputSurface outputSurface;
     private List<GlFilter> filters;
 
-    private FloatBuffer triangleVertices;
-    private final float[] triangleVerticesData = {
-        // X, Y, Z, U, V
-        -1.0f, -1.0f, 0, 0.f, 0.f,
-        1.0f, -1.0f, 0, 1.f, 0.f,
-        -1.0f, 1.0f, 0, 0.f, 1.f,
-        1.0f, 1.0f, 0, 1.f, 1.f,
-        };
-
     private float[] mvpMatrix = new float[16];
-    private float[] stMatrix = new float[16];
-
-    private int glProgram;
-    private int mvpMatrixHandle;
-    private int uStMatrixHandle;
-    private int aPositionHandle;
-    private int aTextureHandle;
 
     /**
-     * Create an instance of GlVideoRenderer
+     * Create an instance of GlVideoRenderer. If filter list has a {@link GlFrameRenderFilter}, that filter
+     * will be used to render video frames. Otherwise, default {@link ScaleToFitGlFrameRenderFilter}
+     * will be used at lowest Z level to render video frames.
      * @param filters optional list of OpenGL filters to applied to output video frames
      */
     public GlVideoRenderer(@Nullable List<GlFilter> filters) {
-        this.filters = filters == null ? Collections.<GlFilter>emptyList() : filters;
+        this.filters = new ArrayList<>();
+        if (filters == null) {
+            this.filters.add(new ScaleToFitGlFrameRenderFilter());
+            return;
+        }
 
-        triangleVertices = ByteBuffer.allocateDirect(
-            triangleVerticesData.length * FLOAT_SIZE_BYTES)
-                                     .order(ByteOrder.nativeOrder()).asFloatBuffer();
-        triangleVertices.put(triangleVerticesData).position(0);
-        Matrix.setIdentityM(stMatrix, 0);
+        boolean hasFrameRenderFilter = false;
+        for (GlFilter filter : filters) {
+            if (filter instanceof GlFrameRenderFilter) {
+                hasFrameRenderFilter = true;
+                break;
+            }
+        }
+        if (!hasFrameRenderFilter) {
+            // if client provided filters don't have a frame render filter, insert default frame filter
+            this.filters.add(new ScaleToFitGlFrameRenderFilter());
+        }
+        this.filters.addAll(filters);
     }
 
     @Override
@@ -130,18 +102,13 @@ public class GlVideoRenderer implements Renderer {
         this.outputSurface = new VideoRenderOutputSurface(outputSurface);
 
         inputSurface = new VideoRenderInputSurface();
-        initGl();
         initMvpMatrix(rotation, aspectRatio);
 
         for (GlFilter filter : filters) {
             filter.init(Arrays.copyOf(mvpMatrix, mvpMatrix.length), 0);
-        }
-
-        // scale video frames to fit the projection
-        if (rotation == 90 || rotation == 270) {
-            Matrix.scaleM(mvpMatrix, 0, 1, aspectRatio, 1);
-        } else {
-            Matrix.scaleM(mvpMatrix, 0, aspectRatio, 1, 1);
+            if (filter instanceof GlFrameRenderFilter) {
+                ((GlFrameRenderFilter) filter).initInputFrameTexture(inputSurface.getTextureId(), inputSurface.getTransformMatrix());
+            }
         }
     }
 
@@ -174,63 +141,11 @@ public class GlVideoRenderer implements Renderer {
     }
 
     /**
-     * Initializes GL state.  Call this after the EGL surface has been created and made current.
-     */
-    private void initGl() {
-        glProgram = GlRenderUtils.createProgram(VERTEX_SHADER, FRAGMENT_SHADER);
-        if (glProgram == 0) {
-            throw new RuntimeException("failed creating glProgram");
-        }
-        aPositionHandle = GLES20.glGetAttribLocation(glProgram, "aPosition");
-        GlRenderUtils.checkGlError("glGetAttribLocation aPosition");
-        if (aPositionHandle == -1) {
-            throw new RuntimeException("Could not get attrib location for aPosition");
-        }
-        aTextureHandle = GLES20.glGetAttribLocation(glProgram, "aTextureCoord");
-        GlRenderUtils.checkGlError("glGetAttribLocation aTextureCoord");
-        if (aTextureHandle == -1) {
-            throw new RuntimeException("Could not get attrib location for aTextureCoord");
-        }
-        mvpMatrixHandle = GLES20.glGetUniformLocation(glProgram, "uMVPMatrix");
-        GlRenderUtils.checkGlError("glGetUniformLocation uMVPMatrix");
-        if (mvpMatrixHandle == -1) {
-            throw new RuntimeException("Could not get attrib location for uMVPMatrix");
-        }
-        uStMatrixHandle = GLES20.glGetUniformLocation(glProgram, "uSTMatrix");
-        GlRenderUtils.checkGlError("glGetUniformLocation uSTMatrix");
-        if (uStMatrixHandle == -1) {
-            throw new RuntimeException("Could not get attrib location for uSTMatrix");
-        }
-    }
-
-    /**
      * Draws the data from SurfaceTexture onto the current EGL surface.
      */
     private void drawFrame(long presentationTimeNs) {
-        GlRenderUtils.checkGlError("onDrawFrame start");
-        inputSurface.getTransformMatrix(stMatrix);
         GLES20.glClearColor(0.0f, 1.0f, 0.0f, 1.0f);
         GLES20.glClear(GLES20.GL_DEPTH_BUFFER_BIT | GLES20.GL_COLOR_BUFFER_BIT);
-        GLES20.glUseProgram(glProgram);
-        GlRenderUtils.checkGlError("glUseProgram");
-        GLES20.glActiveTexture(GLES20.GL_TEXTURE0);
-        GLES20.glBindTexture(GLES11Ext.GL_TEXTURE_EXTERNAL_OES, inputSurface.getTextureId());
-        triangleVertices.position(TRIANGLE_VERTICES_DATA_POS_OFFSET);
-        GLES20.glVertexAttribPointer(aPositionHandle, 3, GLES20.GL_FLOAT, false,
-                                     TRIANGLE_VERTICES_DATA_STRIDE_BYTES, triangleVertices);
-        GlRenderUtils.checkGlError("glVertexAttribPointer maPosition");
-        GLES20.glEnableVertexAttribArray(aPositionHandle);
-        GlRenderUtils.checkGlError("glEnableVertexAttribArray aPositionHandle");
-        triangleVertices.position(TRIANGLE_VERTICES_DATA_UV_OFFSET);
-        GLES20.glVertexAttribPointer(aTextureHandle, 2, GLES20.GL_FLOAT, false,
-                                     TRIANGLE_VERTICES_DATA_STRIDE_BYTES, triangleVertices);
-        GlRenderUtils.checkGlError("glVertexAttribPointer aTextureHandle");
-        GLES20.glEnableVertexAttribArray(aTextureHandle);
-        GlRenderUtils.checkGlError("glEnableVertexAttribArray aTextureHandle");
-        GLES20.glUniformMatrix4fv(mvpMatrixHandle, 1, false, mvpMatrix, 0);
-        GLES20.glUniformMatrix4fv(uStMatrixHandle, 1, false, stMatrix, 0);
-        GLES20.glDrawArrays(GLES20.GL_TRIANGLE_STRIP, 0, 4);
-        GlRenderUtils.checkGlError("glDrawArrays");
 
         for (GlFilter filter : filters) {
             filter.apply(presentationTimeNs);
